# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/salary/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salary"
                ],
                "summary": "创建薪资信息",
                "parameters": [
                    {
                        "description": "薪资信息",
                        "name": "salary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SalaryCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary/delete/{salary_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salary"
                ],
                "summary": "删除薪资信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "薪资ID",
                        "name": "salary_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salary"
                ],
                "summary": "修改薪资信息",
                "parameters": [
                    {
                        "description": "薪资信息",
                        "name": "salary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SalaryEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary/query/all": {
            "get": {
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Salary",
                    "Salary"
                ],
                "summary": "薪资列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary/query/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Salary",
                    "Salary"
                ],
                "summary": "薪资列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary_record/get_salary_record_is_pay_by_id/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SalaryRecord"
                ],
                "summary": "根据ID查询薪资发放记录是否已发放",
                "parameters": [
                    {
                        "type": "string",
                        "description": "薪资发放记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary_record/pay_salary_record_by_id/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SalaryRecord"
                ],
                "summary": "根据ID发放薪资",
                "parameters": [
                    {
                        "type": "string",
                        "description": "薪资发放记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/salary_record/query/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SalaryRecord"
                ],
                "summary": "根据员工ID查询薪资发放记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/salary_record/query_history/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "SalaryRecord",
                    "SalaryRecord"
                ],
                "summary": "薪资历史记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息"
                },
                "status": {
                    "description": "状态码",
                    "type": "boolean"
                }
            }
        },
...
        "model.SalaryCreateDTO": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "integer"
                },
                "bonus": {
                    "type": "integer"
                },
                "commission": {
                    "type": "integer"
                },
                "fund": {
                    "type": "integer"
                },
                "other": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "string"
                },
                "staff_name": {
                    "type": "string"
                },
                "subsidy": {
                    "type": "integer"
                }
            }
        },
        "model.SalaryEditDTO": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "integer"
                },
                "bonus": {
                    "type": "integer"
                },
                "commission": {
                    "type": "integer"
                },
                "fund": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "other": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "string"
                },
                "staff_name": {
                    "type": "string"
                },
                "subsidy": {
                    "type": "integer"
                }
            }
        },
...
    }
}